+++ vision_back/src/main/java/com/vision_back/vision_back/entity/dto/StatusDto.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 
 public class StatusDto {
-    private Long statusId;
-    private Long statusCode;
+    private Integer statusId;
+    private Integer statusCode;
     private String statusDescription; 
 }
